+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Changelog
 
+## [Next version](https://github.com/sdss/sdsstools/compare/0.4.2...HEAD)
+
+
 ## [0.4.2](https://github.com/sdss/sdsstools/compare/0.4.1...0.4.2) - 2020-08-19
 
 - Better support for asyncio exception handling.