+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [Next version](https://github.com/sdss/sdsstools/compare/0.4.2...HEAD)
 
 - Added `Timer` context manager utility.
+- Calling `SDSSLogger.set_level` now also sets the level of any warnings handler already present.
 
 
 ## [0.4.2](https://github.com/sdss/sdsstools/compare/0.4.1...0.4.2) - 2020-08-19