+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## [Next version](https://github.com/sdss/sdsstools/compare/0.4.2...HEAD)
+## [0.4.3](https://github.com/sdss/sdsstools/compare/0.4.2...0.4.3)
 
 - Added `Timer` context manager utility.
 - Calling `SDSSLogger.set_level` now also sets the level of any warnings handler already present.